Documents/EULER PW/requirements.txt
# Dependencias principales de Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Base de datos MySQL
PyMySQL==1.1.0
mysql-connector-python==8.2.0

# Base de datos PostgreSQL (opcional)
psycopg2-binary==2.9.9

# Variables de entorno
python-dotenv==1.0.0

# Utilidades adicionales
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0

# Para desarrollo (opcional)
python-dateutil==2.8.2
six==1.16.0

# Dependencias de SQLAlchemy
SQLAlchemy==2.0.23
typing-extensions==4.8.0
greenlet==3.0.1

# Para migraciones de base de datos
alembic==1.12.1
Mako==1.3.0